requirements.txt
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
PyYAML==6.0.1
reportlab==4.1.0
